Print both 2D and 3D MD temperatures
Previously peakfinder would only display the 3D MD temperature. It is possible to not correctly set the piston velocity and so the reported 3D temperature could be unreliable. Peakfinder will now report the 2D temperature alongside the 3D, so that it is easier to notice an error in setting piston velocity.