Change G^2 algorithm
G^2 algorithm previously would set upper and lower bounds for k,
then use those to populate the position estimate lists.

It is now changed so that it finds all the positions with G^2
under a certain value. This is done by taking the sqrt of the
maximum G^2, using that as the upper and lower bounds of k, then
removing any positions with a G^2 greater than the maximum G^2
value.