Change method of integration
Previously, the integrated intensity was a simple addition of intensities around the volume of the k-space box.

Now, a k volume element is calculated and the integration take place by multiplying the ontensity by the volume element. The code figures out which intensities are in the bulk, on the edges, on a surface, or on a corner; it then multiplies by an appropriate amount of volume depending on the location of the intnesity point. This is to ensure that the integrated intensity only adds contributions from within the defined volume.