Fix bug in finding accurate peak breadths
If the resolution of the function that finds the accurate peak
breadths is too high, there are occasions when the peak maximum
and the neighbouring intensity point have exactly the same
intensity. Previously the condition for finding the breadth was
if the difference in intensities is <= 0.

This has now been fixed so that instead the condition is a
difference < 0.