Start using test driven development
A new method of developing hasd been identified which will allow
more thorough testing of the code, future-proofing it.

The code now has an input file (which can be iterated easily).
The new features are implemented in the "newmodule.py" file. There
are also tests for each of these new features in the
"testpeakfinder.py" file.

The code will start to be structured as follows:
Whenever a new feature is to be implemented, it is to be broken
into smaller tasks, called unit functions. These unit functions
are strung together into composite functions. The composite
function will represent a large operation such as estimating Debye
temperature.